source/requirements.txt
aiohttp==3.8.4
aioredis==1.3.1
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.1
Automat==22.10.0
autopep8==2.0.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.4
channels-redis==4.0.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap4==2022.1
cryptography==39.0.0
daphne==3.0.2
deepgram-sdk==2.3.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.7
django-bootstrap4==22.3
django-braces==1.15.0
django-cors-headers==3.13.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-forms-bootstrap==3.1.0
django-icons==22.1
django-oauth-toolkit==2.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
ecdsa==0.18.0
factory-boy==3.2.1
Faker==16.7.0
frozenlist==1.3.3
hashring==1.5.1
hiredis==2.2.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
MarkupSafe==2.1.2
msgpack==1.0.4
multidict==6.0.4
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pygraphviz==1.10
PyJWT==2.6.0
pymemcache==4.0.0
pyOpenSSL==23.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-magic==0.4.27
python-memcached==1.59
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
redis==4.4.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
simplejson==3.18.3
six==1.16.0
social-auth-app-django==5.1.0
social-auth-core==4.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
style==1.1.0
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.4.0
tzdata==2022.6
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
websockets==10.4
wrapt==1.15.0
yarl==1.8.2
zope.interface==5.5.2
